+++ entity/src/epilot/models/operations/deleteschema.py
@@ -3,19 +3,23 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from typing import Optional
 
 
 @dataclasses.dataclass
 class DeleteSchemaRequest:
+    slug: str = dataclasses.field(metadata={'path_param': { 'field_name': 'slug', 'style': 'simple', 'explode': False }})
+    r"""Entity Type"""
     
-    slug: str = dataclasses.field(metadata={'path_param': { 'field_name': 'slug', 'style': 'simple', 'explode': False }})  
-    
+
+
 
 @dataclasses.dataclass
 class DeleteSchemaResponse:
+    content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
+    status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     
-    content_type: str = dataclasses.field()  
-    status_code: int = dataclasses.field()  
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)  
-    
+